+++ src/polygon/mod.rs
@@ -1,0 +1,3 @@
+mod polygon2;
+
+pub use polygon2::{Polygon2, PolygonBasicValidity, PolygonSelfIntersection};