+++ src/predicates/mod.rs
@@ -3,10 +3,12 @@ mod incircle;
 mod insphere;
 mod orient2d;
 mod orient3d;
+mod polygon2;
 mod sign;
 
 pub use incircle::incircle;
 pub use insphere::insphere;
 pub use orient2d::orient2d;
 pub use orient3d::orient3d;
+pub(crate) use polygon2::polygon_orientation;
 pub use sign::{Orientation, Sign};